# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VisitInput {
  /** @minLength 8 */
  visitorId: string;
  path: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  screenSize?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  language?: string | null;
}

export interface Visit {
  id: number;
  visitorId: string;
  ipAddress: string;
  path: string;
  timestamp: string;
  device: string;
  browser: string;
  os?: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  screenSize?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  latitude?: string | null;
  /** @nullable */
  longitude?: string | null;
  /** @nullable */
  isp?: string | null;
}

export interface Visitor {
  visitorId: string;
  ipAddress: string;
  firstSeen: string;
  lastSeen: string;
  visits: number;
  device: string;
  browser: string;
  os?: string;
  location?: string;
  isBlocked?: boolean;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  latitude?: string | null;
  /** @nullable */
  longitude?: string | null;
  /** @nullable */
  isp?: string | null;
}

export interface VisitorDetail {
  visitor: Visitor;
  activity: Visit[];
}

export type DashboardDailyVisitsItem = {
  date: string;
  visits: number;
};

export interface Dashboard {
  totalVisits: number;
  uniqueVisitors: number;
  repeatVisitors: number;
  blockedIps: number;
  recentActivity: Visit[];
  dailyVisits: DashboardDailyVisitsItem[];
}

export interface BlockInput {
  /** @minLength 3 */
  ip: string;
}

export interface Block {
  ip: string;
  createdAt: string;
}

export type ListVisitorsParams = {
search?: string;
};

export type RemoveBlockParams = {
ip: string;
};