# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardDailyVisitsItem } from './dashboardDailyVisitsItem';
import type { Visit } from './visit';

export interface Dashboard {
  totalVisits: number;
  uniqueVisitors: number;
  repeatVisitors: number;
  blockedIps: number;
  recentActivity: Visit[];
  dailyVisits: DashboardDailyVisitsItem[];
}