# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  visitorId: string;
  ipAddress: string;
  path: string;
  timestamp: Date;
  device: string;
  browser: string;
  os?: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  screenSize?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  latitude?: string | null;
  /** @nullable */
  longitude?: string | null;
  /** @nullable */
  isp?: string | null;
}