# lib/api-zod/src/generated/types/visitInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisitInput {
  /** @minLength 8 */
  visitorId: string;
  path: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  screenSize?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  language?: string | null;
}