# lib/api-zod/src/generated/types/visitor.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visitor {
  visitorId: string;
  ipAddress: string;
  firstSeen: Date;
  lastSeen: Date;
  visits: number;
  device: string;
  browser: string;
  os?: string;
  location?: string;
  isBlocked?: boolean;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  latitude?: string | null;
  /** @nullable */
  longitude?: string | null;
  /** @nullable */
  isp?: string | null;
}