# lib/api-zod/src/generated/types/visitorDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Visit } from './visit';
import type { Visitor } from './visitor';

export interface VisitorDetail {
  visitor: Visitor;
  activity: Visit[];
}